refactor: move typewriter hook into lib
