+++ lib/useTypewriter.ts
